processor/requirements.txt
pyspark==3.5.0
python-dotenv